Refactor command matching into own module
